extract compression-stream into its own files
